requirements.txt
pandas
fireducks
scikit-learn
matplotlib
python-dotenv
pyarrow
